Superoptimizer audit: fix 12+ issues, add x86-64 instruction selection
P0 fixes (critical compile-time/bugs):
- S1: count_nodes() O(N) per simulation → O(1) node_count field
- S2: Var(String) → Var(u32) with StringInterner (5-10x alloc reduction)
- S9: Absorb rewrite bug - now checks lhs == inner_lhs equality
- S10: Reuse MctsSuperoptimizer across expressions (avoid CPUID per expr)
- S13: Cache TargetConfig in SuperoptPass (avoid CPUID per expr)

P1 fixes (high impact):
- S3: apply_action_recursive depth cutoff (prevent exponential clone storms)
- S4: collect_variables uses HashSet for O(1) dedup
- S6: Deduplicate children in expand_node using hash-based seen set

P2/P3 fixes:
- S11: hot path uses hotpath() instead of maximum() (25x over-budget)
- S26: Fixpoint cap reduced 32→8 (most rewrites converge in 2-3 iters)
- S27: OpcodeTable::standard() moved outside loop in is_well_formed

Architecture improvements:
- Added LEA, INC, DEC rewrite rules to MCTS (LeaCombine, IncDec, LeaMulSmallConst)
- Cost model now recognizes LEA patterns: (x<<k)+y → single LEA instruction
- Added LEA/INC/DEC uop entries to all 7 microarchitecture port maps
- Flatten step emits 'lea', 'inc', 'dec' for better cost estimation
- LeaMulSmallConst: x*3→(x<<1)+x, x*5→(x<<2)+x, x*9→(x<<3)+x, x*7→(x<<3)-x

Bug fix (pre-existing):
- phase3_jit.rs: is_power_of_two() on i64 → cast to u64 first

+++ src/jit/phase3_jit.rs
@@ -1509,7 +1509,7 @@ fn cse_optimize(instrs: &mut Vec<Instr>) {
 /// Returns (magic_constant, post_shift) or None if IDIV should be used instead.
 #[allow(dead_code)]
 fn compute_div_magic(d: i64) -> Option<(i64, u8)> {
-    if d <= 0 || d.is_power_of_two() || d == 1 {
+    if d <= 0 || (d as u64).is_power_of_two() || d == 1 {
         return None;
     }
     let d = d as u64;
